[ES-2388] Fixed the issue where the UI was stuck when signup client was onboarded with multiple authFactors..
